pyproject.toml
[tool.poetry]
name = "rivergen"
version = "1.2.3"
description = "Procedural Road/River Segment Generator for NN Training"
requires = { python = ">=3.10" }
readme = "README.md"
license = "MIT" 
authors = [
    "Niklas Paulig <niklas.paulig@tu-dresden.de>"
]
keywords = ["procedural generation", "road", "river", "NN training"]
homepage = "https://github.com/nikpau/segGen"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[tool.poetry.dependencies]
python = ">=3.10"
numpy = ">=1.19"
attrs = ">=22.0.0"
matplotlib = ">=3.5.3"
pyyaml = "*"

[tool.poetry.scripts]
rivergen = "rivergen.__main__:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"